requirements.txt
ultralytics
rasterio
geopandas
torch
torchvision
tqdm
numpy
Pillow
shapely